+++ app/renderer/components/preferences/payment/enabledContent.js
@@ -14,7 +14,7 @@ const {changeSetting} = require('../../../lib/settingsUtil')
 // components
 const ImmutableComponent = require('../../immutableComponent')
 const Button = require('../../../../../js/components/button')
-const {FormTextbox} = require('../../textbox')
+const {FormTextbox} = require('../../common/textbox')
 const {FormDropdown} = require('../../common/dropdown')
 const {SettingsList, SettingItem} = require('../../settings')
 const LedgerTable = require('./ledgerTable')
